+++ canvas_sdk/v1/data/__init__.py
@@ -180,6 +180,7 @@ from .questionnaire import (
 )
 from .reason_for_visit import ReasonForVisitSettingCoding
 from .referral import Referral, ReferralReport, ReferralReportCoding, ReferralReview
+from .refill_request import RefillRequest, RefillRequestCoding
 from .service_provider import ServiceProvider
 from .snapshot import Snapshot, SnapshotImage
 from .specialty_report_template import (
@@ -388,6 +389,8 @@ __all__ = __exports__ = (
     "ReferralReport",
     "ReferralReportCoding",
     "ReferralReview",
+    "RefillRequest",
+    "RefillRequestCoding",
     "ResponseOption",
     "ResponseOptionSet",
     "ServiceProvider",